+++ src/App.tsx
@@ -15,6 +15,7 @@ import Farm from "views/Farm";
 import Home from "views/Home";
 import Addresses from "views/Addresses";
 import Help from "views/Help";
+import Vote from "views/Vote";
 import DaoInfo from "views/DaoInfo";
 
 const App: React.FC = () => {
@@ -39,6 +40,7 @@ const App: React.FC = () => {
             <Route path="/farm" element={<Farm />} />
             <Route path="/addresses" element={<Addresses />} />
             <Route path="/help" element={<Help />} />
+            <Route path="/vote" element={<Vote />} />
             <Route path="/daoinfo" element={<DaoInfo />} />
           </Routes>
         </div>